-New method using netlist to read packets on DVD device by sectors of 2048 bytes. -Some modifications in netlist to make it work with DVD. -Fixed � bug in mpeg_system.c
We use INPUT_READ_ONCE many times in the input code. I think that
we should replace it with a field in input_thread_t to give more
flexibility to modules. I have done it in DVD plugin but it can't work
if input.c is limited by INPUT_READ_ONCE.

The DVD method is slower than input_ps ; but there are things to
improve.
CSS descrambling is broken.