-Changes in the way stream size in DVD mode is calculated. It is no longer such an ugly kludge and is much more accurate.
-Changes in DVDSeek so that the interface slider is in 0 when we are at
the beginning of the strem.

-Try to detect the beginning of the movie in the middle of the title. It
works for most of the DVDs but will begin in the middle of the movie
with some others.

-Calculate only the first title key in css mode because it might take a
long time with some title. The search stops now when we have a key once
whereas it did stop only when we had it twice before. It should not
cause problems.