Added VCD input plugin for Linux :
The Linux specific stuff is located in linux_cdrom_tools files.
For the time being, it only has title (tracks). Chapters will come if I
ever can read .vcd files.
I've modified Gnome and GTK+, not the others.
I've noticed a big bug, it segfaults at the end of a title. I will
have a look on this tomorrow.
Please be indulgent, it's my first peace of code ;-)