Support for libdvdread.
Libdvdread files are in extras/libdvdread. They have not been changed from
those of libdvdread 0.9.2. The changes are in the subdirectory videolan.
They basically consist of a new function in the API that does a readv()
instead fo a read().

The library is statically linked with libdvdcss from extras/libdvdcss, and
the whole is statically linked with the plugin dvdread. Dvdread can only be
compiled as a builtin module for the time being, I think (which means I
haven't tried to compile it as a plugin).

Basically, input_dvdread.c is a copy and paste from input_dvd.c with a few
changes to use libdvdread calls. It is still in very alpha stage, but it
should offer the same features as the current DVD input.

It can be called with either the syntax: "dvdread:/dev/dvd"
or "--input dvdread /dev/dvd". The command line switches (-t, -T, -c, -u)
should work too.