* New GTK configuration interface. This interface is generated dynamically   at run-time from the configuration data contained in each plugin. * config_PutPszVariable and config_GetPszVariable are now thread safe. * Few modifications/optimizations to the config module. * Got rid of all the _VAR macros in config.h
This is the first version of the GTK configuration interface. There are still
a few raw corners and a fews things left to do, but it basically does what
I wanted.
Constructive comments like it sucks or it smells rotten fish are of course
very welcome. And if there are GTK gurus out there, please don't hesitate to
add your touch.

I also started to add usefull comments to the configuration options but I'm not
really good at this kind of thing so if people want to help, they can finish
the job. For this look at all the *_LONGTEXT macros at the top of main.c

(PS: Translations will also be appreciated but not right now as the interface
will likely change again in the few coming days)