* First try for IPv6 support on Win32. This hasn't been tested so it   might not work! To compile this stuff you need a mingw compiler with   w32api-1.3.
The win32 code uses the getaddrinfo function because inet_pton is
supposed to be deprecated and not supported by Microsoft. gethostbyname2
isn't supported as well. Maybe we could try to unify the ipv6 plugin and
have all the ports use getaddrinfo ? (I didn't dare to do it because I
don't want to break anything before a release ;-)