* skins/controls/button.cpp: Fixed a nasty bug.  * skins/controls/generic.h: Some public fields are now protected  * skins/src/window.cpp:     - Added a few debug messages     - Invisible controls don't receive events anymore
Note: There are still a few bugs in the code handling the controls and
their different states, and I think they will be difficult to fix with
the current code structure. A rewrite of this part of code could be a
good idea (perhaps using states and transitions between states, like
Zinf does?).