dvbsub.c: DVB subtitles decoder (ETSI EN 300 743). There is still a lot of           things to write to implement competely the standard but this decoder           will work in most of the pictures subtitles cases. Text subtitles           decoding is not implemented at all.           This decoder has been validated with CAVENA encoders after few           remarks of their developpers.
include/*: addded a p_spuinfo field in the decider_fifo_t and the
           es_descriptor_t

codecs.h: added a dvb_spuinfo_t structure to identify the id of the selected
          track.

modules/demux/mpeg: added the support of the dvb subtitles track. Need last
                    libdvbpsi (0x59 descriptor) to compile. The demuxer
                    creates n fake ES as each ES carries up to 256 subtitles
                    tracks.


This work is part of the work done by Anevia for the "Gran Theatro del
Barcelona".