* Implement INPUT_GET_SUBDELAY and INPUT_SET_SUBDELAY
* Allow to modify subtitle delay on the fly and implement hotkeys 
  (h and j)

* When framerate is available with subrip, use sub-fps to make corrections
  (sub-fps default value is now 0, if left to this value, 25fps will be
   assumed in microdvd mode and no correction will be applied in subrip)

Main problem is that when seeking in the subtitles, you easily get
several of them displayed. Maybe we could add a flag to subpicture_t,
like SUBTITLE_PICTURE to say that only one of them must be displayed at
a time, but that other subpictures (OSD, ...) can remain present on the
screen at the same time.