Same fix for non-constant string.
The playlist works fine again (with ASCII/Latin locale charset).
When in UTF8, it might be a good idea to check for invalid bytes
sequences and replace them with question mark - at the moment
wxWidgets still leave the whole text blank which result in blank
playlist item in the playlist.