* Fix for the disabled menu's under Tiger.
Apparently menu's got smarter in Tiger. As a workaround for the old version, you should first open the Help menu and then drag to the other menu's.

This still isn't the correct way to fix it. We should attach a callback function to the autogenerated menu's and have the callback function update the menu. (but this will work for now).