Don't set LC_NUMERIC in the wxWidgets plugin.
It's already set to C globally because there are a bunch of things that
assume US-style float numbers formatting (config file, RTSP VoD server
to name a few) -- see SetLanguage() in libvlc.

If you want to get a float from the user, you can look for the comma
and replace it with a dot before parsing. However keep in mind that the
stream output parser already uses the comma to separate options; that's
yet another reason to use the dot in any case.

Also, I don't know if some remote locales use neither dot nor comma as
their decimal point.