* added the needed libraries for UPnP-support on OSX  - CyberLink C++ UPnP 1.7.1  - libexpat 1.95.8
Note that Cyberlink compiles fine, but must not access its own install script, so we need to run "chmod 755" on this script first. 
The general patch was created by courmisch. Thanks again!

The additional osx-patch is a really ugly hack which disables libiconv to let our upnp-plugin compile. I'll further investigate this in the next couple of days (see ticket #367).