* implemented the controls for the logo and time filters (refs #350) * complete l10n support
The usage of the logo-filter doesn't work here, but that might be because of my strange PNG-files. I'll investigate this tomorrow.

It would be nice if some of the OSX-devs and testers could check whether everything works as expected, whether you are able crash the entire thingie, etc.
Note that on-the-fly activation of filters doesn't work (this needs the vout-rework scheduled for 0.8.5).