As a work-around, accept both decimal separators when... deserializing floats from configuration (fixes #589)
I know this is NOT the “right” way to fix this.
The “right” way would be to remove floats from the configuration...
but there are approximately one float parameter in the core and 46 ones
in our current plugins set, none of which I use myself. So I am
defintely not going to rewrite all^H^H^Hany of them.

Of course, it would be cleaner to be “strict”  in what we write to
the configuration file, than “liberal” in what we accept... but IMHO,
it is way easier and more robust to be liberal in what we accept in
this particular case (ie. accepting both decimal separators is trivial,
while serializing float numbers manually is complex).