Run FontBuilder thread after Freetype has been initialized and no error paths are can be taken. This should solve the crash reported by several people with the video title (or meta-name) appearing at start of the movie.
The freetype module starts another thread called FontBuilder and depending on timing this creates a crash or works allright. The crash is triggered by the font not being found. This patch tries to load the font first before the FontBuilder thread is started.

I was unable to reproduce the crash with the patch applied. Without the patch the crash is reproducable.

As discussed on the mailinglist the FontBuilder thread doesn't depend on the freetype initialisation and rearanging the order should do no harm.