Revert "prevent #1312 from occuring (refs #1312)"
This reverts commit 69d572ca8a6bc6f9de85ed471fbb30e430fb0fe2.
It's not an issue with sout-keep alone. The issue shows up if using
both sout-keep and #gather, and there is no easy way for a stream
object to know what its upstream stream (if any) is. Hence, let users
shoot themselves in the feet if they want, and otherwise assume that
people don't use #gather _in_front_ of #display.