stream: Add a new method for buffering access: A*Immediate method.
It is much more efficient regarding latency as it doesn't bufferize more than needed, and let the module access take care of that eventually.

Enable with --use-stream-immediate. We may want to default it.

Note: --use-stream-immediate will be only effective on access that don't provide pf_block() for now. This is because I didn't benchmark against the Block method.

Here the gain that I did measure is about 200ms (less latency) when using the http access on a loopback.