Revert "Remove VLC_PRIVATE_API"
This reverts commit 96cc9c261db1c616efa2fbc0aea21bc70b1e46fc.

This commit is reverted because we are using a regexp to determine the list of exported symbols in libvlc. However some private symbols start with libvlc_, this produces (here on Mac OS X) a nmedit error about symbols not being found while exported. A solution is to externalize those private symbols, hence using VLC_PUBLIC_API.

Feel free to come up with a better solution.