mkv: Find and load a 2nd Seek head and use the information contained in it.
This fixes issues with files which had their Cues referenced from this 2nd Seekhead at the end of the file. These files were previously treated as "broken".
A maximum of 10 seek heads is allowed in our module, to prevent recursion. The module is much more flexible now, but not that some Master elements that are allowed to be present multiple times are still handled only once (KaxInfo, KaxTracks, KaxTags).

With these change, most mkv files produced with mkvmerge will produce a significant amount of debug. This is because these files have a load of Junk tags in their eof Seekhead that our parser is warning about. After consulting with robux, it was determined that this was a bug in mkvmerge and a ticket has been filed ( https://www.bunkus.org/bugzilla/show_bug.cgi?id=290 ).

This fixes #1236