mkv: lower priority of mkv module to 0
I've had it with this module. There is no support, let alone development, from the matroska folks anymore. I prefer to use ffmpeg which seems to be at least less crash prone and having more of a future. I'm not sure which features we are loosing due to this atm. Likely some meta info, probably ordered chapters and definetly the DVD menu capabilities.