variables: VAR_CLASS != VAR_TYPE
This should fix a memleak with variable strings, and several other problems. Note that a CLASS detemines the basic type, and that TYPE is only a specific set within the CLASS. The variable types: VLC_VAR_KEY and VLC_VAR_INTEGER are both of CLASS integer. and thus compatible, when it comes to inheritence/init/free etc.