commandline parsing: Fix the macfreetv crash
Basically, somewhat of a "duh" now that we have optional_arguments on OSX.
It is highly annoying however that getopt_long behaviour differs so much
between gnu and BSD implementations, and then is even BUGGY on darwin.
'--option=value' is uniformly supported  and '--option value' is a GNU getopt thing.

imperfect world etc...