update: allow update_Download  to accept: directory, filename, NULL
directory: accepts any existing directory and will use the filename component of the update URL (old behaviour). This option no longer requires you to provide a string that ends with the pathsep.

filename: write directly to this filename. Existing files will be overwritten.

NULL: will write to filename component of the update URL. likely in the CWD.