Don't start the file logger when syslog is specified.
If --syslog is specified on the commandline then vlc loads the logger in file mode first, because --file-logging is the default. This is counter intuitive for users, they expect *only* syslog mode to be active. Note that the logger module can only have one mode active at the sametime, either html, text or syslog.