DRM Video decryption
The attached patch adds support for decrypting and playing drm
protected video, such as those from iTunes movie store.

The patch was created against the 0.9.5 source tarball, and was
compiled and tested on OSX against some (legitimately purchased)
iTunes purchased songs and movies.

Just like the existing audio support it requires decryption keys to be
deposited in ~/.drms/ and extracting the keys is left as an exercise
for the user.

[Commit message edited by Laurent Aimar]

Signed-off-by: Iain Wade <iwade@optusnet.com.au>
Signed-off-by: Laurent Aimar <fenrir@videolan.org>