Thread cancellation on WinCE
This patch adds support for thread cancellation on Windows CE. It
emulates missing functions such as WaitForSingleObjectEx, by creating
a cancellation event handle, and forwarding the wait functions to wait
on both the original handles and the cancellation event. The thread can
then be cancelled by triggering the event from another thread.

Signed-off-by: Rémi Denis-Courmont <rdenis@simphalempin.com>