mux/ogg: Don't use granulepos = UINT64_MAX on EOS page
It isn't legal for granulepos to be UINT64_MAX on an EOS page since
this special value indicates that no packets finish on the page, yet
an EOS page signifies that the final packet does finish on the page.

It seems that there is a convention for repeating the last used
granulepos in the EOS page per logical stream.  This then allows
some extra file size hints (duration = last gp - first gp) [1]

[1] This is flawed -- EOS may occur anywhere, however this is what
some people do.

Signed-off-by: David Flynn <davidf@rd.bbc.co.uk>
Signed-off-by: Laurent Aimar <fenrir@videolan.org>