Added vlc_get_libvlc_object to libvlccore
It is intented to expose the libvlc_int_t (the main vlc object)
to enable extented vlc hacking.. E.g. when you need to make something
with configuration or modules that it is not provided by the PUBLIC API,
you expose the libvlc_int_t as an vlc_object_t and act upon it with
the internal (libvlccore) API...

Signed-off-by: Rémi Denis-Courmont <rdenis@simphalempin.com>

Put back to libvlc rather than libvlccore (a stupid idea of mine).