Support for rendering Kate streams with libtiger
libtiger is a rendering library for Kate streams based on Pango
and Cairo, and supports almost everything Kate streams can do.

There's a bit of wonkiness with rendering on regions as VLC then
blends them onto a subsampled YUV buffer, and colors can bleed
in a not so nice way. I've tried frobbing the blender in blend.c
and I got rid of the chroma bleeding, but at the cost of other
artifacts, so I'm leaving it out.

Also, I can't find a way to get the preferences dialog to call
a callback on my variables with a pointer to the decoder, so
changing the variables will not have effect till VLC is started
again.

Also includes the misc kate fixes from a previous patch, which
was not applied yet, but changes the same code as this patch
(and fixed a possible crash).

Note that I don't have a big endian host to check, so the RGBA
component flipping code is untested.

Signed-off-by: Rémi Denis-Courmont <rdenis@simphalempin.com>