Qt4: systray hides the interface when it is visible regardless of focus
Otherwise, if the interface is hidden behind another window, clicking
the tray icon however many times will request user attention in the task
bar and never actually show the interface. This was confusing and
inconvenient (fixes #2396). This new behavior is consistent with KMail.

It would probably be even better to show the interface on top (instead
of hiding it, then show it on the second click). Unfortunately, while it
does seem possible, I do not know how. For sure, the previous scheme did
not work anyway.