X C Bindings video output proof of concept plugin
This can really only output iamges at this point.
Critical yet-to-be-implemented features:
 - MIT-SHM shared memory support
 - XVideo hardware scaling & chroma conversion
 - keyboard event handling
 - mouse event handling
 - window (manager) event handling
 - cursor blanking
 - window provider submodule (so it can work without the Qt interface)
 - screen capture access module
 - GLX, perhaps?
 - etc

15, 16 and 24-bits color depth work. 32-bits depth was not tested due
to lack of suitable video adapter. 8-bits depth does not work due to a
bug in VLC chroma handling. B&W, grey and 16 color are not supported,
but are probably not worth it - the Xlib output does not support them
either.

Fullscreen should work, but was not tested.